+++ test/test_query2.cpp
@@ -5817,7 +5817,7 @@ TEST(Query_Mixed)
             obj.set(col_mixed, Mixed(i));
         }
         else {
-            obj.set(col_mixed, Mixed(table->begin()->get_key()));
+            obj.set(col_mixed, Mixed(table->begin()->get_link()));
         }
         for (int64_t j = 0; j < 10; j++) {
             ll.add(it->get_key());
@@ -5890,9 +5890,9 @@ TEST(Query_Mixed)
     tv = (table->column<Mixed>(col_any) == table->column<Int>(col_int)).find_all();
     CHECK_EQUAL(tv.size(), 72);
 
-    tv = (origin->column<Mixed>(col_mixed) == Mixed(table->begin()->get_key())).find_all();
+    tv = (origin->column<Mixed>(col_mixed) == Mixed(table->begin()->get_link())).find_all();
     CHECK_EQUAL(tv.size(), 4);
-    tv = (origin->where().links_to(col_mixed, table->begin()->get_key())).find_all();
+    tv = (origin->where().links_to(col_mixed, table->begin()->get_link())).find_all();
     CHECK_EQUAL(tv.size(), 4);
     tv = (origin->link(col_links).column<Mixed>(col_any) > 50).find_all();
     CHECK_EQUAL(tv.size(), 5);